// src/gen/user_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file user.proto (package example, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file user.proto.
 */
export const file_user: GenFile = /*@__PURE__*/
  fileDesc("Cgp1c2VyLnByb3RvEgdleGFtcGxlIrQBCghVc2VyS2V5cxIXCg91c2VyX3B1YmxpY19rZXkYASABKAwSHQoQdXNlcl9wcml2YXRlX2tleRgCIAEoDEgAiAEBEhUKCGRhdGFfa2V5GAMgASgMSAGIAQESFQoIbGlua19rZXkYBCABKAxIAogBARITCgtzdG9yYWdlX2tleRgFIAEoDEITChFfdXNlcl9wcml2YXRlX2tleUILCglfZGF0YV9rZXlCCwoJX2xpbmtfa2V5YgZwcm90bzM");

/**
 * @generated from message example.UserKeys
 */
export type UserKeys = Message<"example.UserKeys"> & {
  /**
   * @generated from field: bytes user_public_key = 1;
   */
  userPublicKey: Uint8Array;

  /**
   * @generated from field: optional bytes user_private_key = 2;
   */
  userPrivateKey?: Uint8Array;

  /**
   * @generated from field: optional bytes data_key = 3;
   */
  dataKey?: Uint8Array;

  /**
   * @generated from field: optional bytes link_key = 4;
   */
  linkKey?: Uint8Array;

  /**
   * @generated from field: bytes storage_key = 5;
   */
  storageKey: Uint8Array;
};

/**
 * Describes the message example.UserKeys.
 * Use `create(UserKeysSchema)` to create a new message.
 */
export const UserKeysSchema: GenMessage<UserKeys> = /*@__PURE__*/
  messageDesc(file_user, 0);